refactor(pty): split TerminalProcess into focused modules
- Extract terminalInput.ts (input chunking, paste handling, submit timing)
- Extract terminalSessionPersistence.ts (session restore and persistence)
- Extract terminalActivityPatterns.ts (pattern compilation, process validation)
- Extract TerminalForensicsBuffer.ts (crash forensics and logging)
- Extract terminalShell.ts (shell detection and defaults)
- Fix agent type resolution in getSoftNewlineSequence to check detectedAgentType/agentId
- Reduce TerminalProcess.ts from 1,700 to 1,350 lines
